Vijith/backend

+++ backend/OpenElective/OpenElective/Controllers/OpenElectivesController.cs
@@ -2,12 +2,14 @@
 using OpenElective.Services.Interfaces;
 using AutoMapper;
 using OpenElective.Models.DTOs.OpenElectives;
-using OpenElective.Models;
+using Microsoft.AspNetCore.Authorization;
+using System.Security.Claims;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
 namespace OpenElective.Controllers
 {
+    [Authorize]
     [Route("api/[controller]")]
     [ApiController]
     public class OpenElectivesController : ControllerBase
@@ -21,6 +23,7 @@ namespace OpenElective.Controllers
             this.mapper = mapper;
         }
         // GET: api/<OpenElectivesController>
+        [Authorize(Roles = "admin")]
         [HttpGet]
         public IActionResult Get()
         {
@@ -37,6 +40,7 @@ namespace OpenElective.Controllers
         }
 
         // GET api/<OpenElectivesController>/5
+        [AllowAnonymous]
         [HttpGet("{id}")]
         public IActionResult Get(Guid id)
         {
@@ -55,6 +59,7 @@ namespace OpenElective.Controllers
         }
 
         // POST api/<OpenElectivesController>
+        [Authorize(Roles = "admin")]
         [HttpPost]
         public IActionResult Post([FromBody] CreateOpenElectiveDTO openElective)
         {
@@ -63,6 +68,12 @@ namespace OpenElective.Controllers
                 if(openElective == null)
                 {
                     return BadRequest(openElective);
+                }
+                var claimsIdentity = User.Identity as ClaimsIdentity;
+                var IdClaim = claimsIdentity.FindFirst(ClaimTypes.Role);
+                if (IdClaim.Value.ToString() != "admin")
+                {
+                    return Forbid();
                 }
                 var OE=mapper.Map<OpenElective.Models.OpenElective>(openElective);
                 OE.Id = Guid.NewGuid();
@@ -83,17 +94,27 @@ namespace OpenElective.Controllers
         }
 
         // PUT api/<OpenElectivesController>/5
+
+        [Authorize(Roles = "admin")]
+
         [HttpPut("{id}")]
-        public IActionResult Put(Guid id, [FromBody] OpenElective.Models.OpenElective openElective)
+        public IActionResult Put(Guid id, [FromBody] UpdateOpenElectiveDTO openElectiveDTO)
         {
             try
             {
-                if (openElective == null)
+                if (openElectiveDTO == null)
                 {
-                    return BadRequest(openElective);
+                    return BadRequest(openElectiveDTO);
                 }
-
-                var updatedOE = openElectiveService.Update(openElective);
+                var claimsIdentity = User.Identity as ClaimsIdentity;
+                var IdClaim = claimsIdentity.FindFirst(ClaimTypes.Role);
+                if (IdClaim.Value.ToString() != "admin")
+                {
+                    return Forbid();
+                }
+                var oe=openElectiveService.Get(id);
+                oe.Name = openElectiveDTO.Name;
+                var updatedOE = openElectiveService.Update(oe);
                 return CreatedAtAction(nameof(Put), updatedOE);
 
             }
@@ -105,17 +126,24 @@ namespace OpenElective.Controllers
         }
 
         // DELETE api/<OpenElectivesController>/5
+        [Authorize(Roles = "admin")]
+
         [HttpDelete("{id}")]
         public IActionResult Delete(Guid id)
         {
             try
             {
+                var claimsIdentity = User.Identity as ClaimsIdentity;
+                var IdClaim = claimsIdentity.FindFirst(ClaimTypes.Role);
+                if (IdClaim.Value.ToString() != "admin")
+                {
+                    return Forbid();
+                }
                 var deleted=openElectiveService.Delete(id);
                 return Ok(deleted);
             }
             catch (Exception)
             {
-
                 throw;
             }
         }
